@@ -1,6 +1,6 @@
 {
 	"name": "skynovel2",
-	"version": "4.18.4",
+	"version": "4.19.0",
 	"publisher": "famibee2",
 	"displayName": "SKYNovel",
 	"description": "It supports development of SKYNovel for VSCode, VSCodeでSKYNovelの開発をサポートする拡張機能",
@@ -9,7 +9,7 @@
 	"homepage": "http://famibee.blog38.fc2.com/",
 	"license": "MIT",
 	"engines": {
-		"vscode": "^1.95.0"
+		"vscode": "^1.96.0"
 	},
 	"categories": [
 		"Programming Languages",
@@ -706,45 +706,49 @@
 		"await-notify": "^1.0.1",
 		"crc-32": "^1.2.2",
 		"fs-extra": "^11.2.0",
-		"npm-check-updates": "^17.1.11",
+		"npm-check-updates": "^17.1.12",
 		"original-fs": "^1.2.0",
-		"pinia": "^2.2.6",
+		"pinia": "^2.3.0",
 		"socket.io": "^4.8.1",
 		"uuid": "^11.0.3",
-		"vee-validate": "^4.14.7",
+		"vee-validate": "^4.15.0",
 		"vscode-languageclient": "^9.0.1",
 		"vue": "^3.5.13",
-		"yup": "^1.4.0"
+		"yup": "^1.6.1"
 	},
 	"devDependencies": {
-		"@types/adm-zip": "^0.5.6",
+		"@ffmpeg-installer/ffmpeg": "^1.1.0",
+		"@types/adm-zip": "^0.5.7",
 		"@types/archiver": "^6.0.3",
 		"@types/bootstrap": "^5.2.10",
+		"@types/fluent-ffmpeg": "^2.1.27",
 		"@types/fs-extra": "^11.0.4",
 		"@types/jest": "^29.5.14",
-		"@types/node": "^22.9.0",
+		"@types/node": "^22.10.2",
 		"@types/power-assert": "^1.5.12",
 		"@types/psd": "^3.4.3",
-		"@types/vscode": "^1.95.0",
+		"@types/sharp": "^0.32.0",
+		"@types/vscode": "^1.96.0",
 		"@types/vscode-webview": "^1.57.5",
-		"@typescript-eslint/eslint-plugin": "^8.14.0",
-		"@typescript-eslint/parser": "^8.14.0",
-		"@vitejs/plugin-vue": "^5.2.0",
+		"@typescript-eslint/eslint-plugin": "^8.18.2",
+		"@typescript-eslint/parser": "^8.18.2",
+		"@vitejs/plugin-vue": "^5.2.1",
 		"@vscode/test-electron": "^2.4.1",
-		"electron": "^33.2.0",
-		"esbuild": "^0.24.0",
+		"electron": "^33.2.1",
+		"esbuild": "^0.24.2",
 		"file-loader": "^6.2.0",
-		"image-size": "^1.1.1",
+		"image-size": "^1.2.0",
 		"p-queue": "^8.0.1",
+		"png2icons": "^2.0.1",
 		"rimraf": "^6.0.1",
-		"terser": "^5.36.0",
+		"terser": "^5.37.0",
 		"ts-loader": "^9.5.1",
-		"typescript": "^5.6.3",
-		"vite": "^5.4.11",
+		"typescript": "^5.7.2",
+		"vite": "^6.0.5",
 		"vscode-languageserver-types": "^3.17.5",
 		"vue-loader": "^17.4.2",
-		"webpack": "^5.96.1",
-		"webpack-cli": "^5.1.4"
+		"webpack": "^5.97.1",
+		"webpack-cli": "^6.0.1"
 	},
 	"optionalDependencies": {
 		"bufferutil": "^4.0.8",
@@ -759,7 +763,7 @@
 		"update": "bun update && bun run build && bun ncu_lsp && ncu -g",
 		"ncu_lsp": "cd server && bun update && bun outdated && cd ..",
 		"postinstall": "cd server && bun i && cd ..",
-		"rebuild": "rimraf package-lock.json && rimraf node_modules && bun pm cache rm && bun i",
+		"rebuild": "rimraf node_modules && bun pm cache rm && bun i",
 		"version:patch": "npm -no-git-tag-version version patch",
 		"version:minor": "npm -no-git-tag-version version minor",
 		"version:major": "npm -no-git-tag-version version major",
