@@ -5,15 +5,18 @@
 	http://opensource.org/licenses/mit-license.php
 ** ***** END LICENSE BLOCK ***** */
 
-const [, , ...aCmd] = process.argv;
+const [, , fld_src='', ...aCmd] = process.argv;
 const minify = aCmd.includes('--minify');
 
 import {exec} from 'child_process';
-import {statSync, writeFile, writeJsonSync, existsSync, copy} from 'fs-extra';
+import {stat} from 'fs/promises';
+import {outputFile, writeJsonSync, pathExistsSync, copy} from 'fs-extra/esm';
 const is_win = process.platform === 'win32';
 import {userInfo} from 'os';
 import {extname} from 'path';
 
+import url from 'url';
+const __filename = url.fileURLToPath(import.meta.url);
 
 type LOG = {inp: string, out: string, iSize: number, oSize: number, err: string};
 const oLog: {[nm: string]: LOG} = {};
@@ -26,7 +29,7 @@ const fnc: (log: LOG, nm: string, str: string)=> Promise<void> = minify
 	? async (log, nm, str)=> {
 		try {
 			const fnTmp = __filename.slice(0, -3) +'_'+ nm +'.txt';
-			await writeFile(fnTmp, str, {encoding: 'utf8'});
+			await outputFile(fnTmp, str, {encoding: 'utf8'});
 
 			str = str.replaceAll('\"', '\\"');
 			await new Promise<void>((re, rj)=> exec(`pyftsubset "${log.inp}" --text-file="${fnTmp}" --layout-features='*' --flavor=woff2 --output-file="${log.out}" --verbose`, async (e, _stdout, stderr)=> {
@@ -38,7 +41,7 @@ const fnc: (log: LOG, nm: string, str: string)=> Promise<void> = minify
 					return;
 				}
 
-				await writeFile(fnTmp, stderr, {encoding: 'utf8'});
+				await outputFile(fnTmp, stderr, {encoding: 'utf8'});
 				if (stderr.includes('Missing glyphs for requested Unicodes:')) log.err += `${nm} 出力警告：フォントファイルに含まれない文字がありました。ログを確認（Missing glyphs ...）して下さい。\n`;
 				re();
 			}));
@@ -49,11 +52,14 @@ const fnc: (log: LOG, nm: string, str: string)=> Promise<void> = minify
 	}
 	: log=> copy(log.inp, log.out);
 
-const o = require('./font.json');
+// const o = require('./font.json');
+// @ts-expect-error
+import o from './font.json' with {type: "json"};
+// import o from './font.json';
 const a = [];
 
 const {username} = userInfo();
-const PATH_PRJ_FONTS = 'core/font';
+const PATH_PRJ_FONTS = `${fld_src}/font`;
 const PATH_USER_FONTS = is_win
 	? `C:/Users/${username}/AppData/Local/Microsoft/Windows/Fonts`
 	: `/Users/${username}/Library/Fonts`;
@@ -68,24 +74,24 @@ for (const [nm, v] of Object.entries(o)) {
 	.replace('::PATH_OS_FONTS::', PATH_OS_FONTS);
 	const out = `doc/prj/script/${nm}${minify ?'.woff2' :extname(inp)}`;
 	const log = oLog[nm] = {inp, out, iSize: 1, oSize: 1, err: ''};
-	if (! existsSync(inp)) {
+	if (! pathExistsSync(inp)) {
 		log.err = `変換失敗です。入力ファイル ${o[nm].inp.slice(20)} が存在するか確認してください`;
 		continue;
 	}
 	a.push(fnc(log, nm, o[nm].txt));
 }
 Promise.allSettled(a)
-.then(()=> {
+.then(async ()=> {
 	for (const [nm, log] of Object.entries(oLog)) {
 		const {inp, out} = log;
 		log.inp = o[nm].inp;	// プライベートな環境値を塗りつぶす
 	//	log.out = // これは存在しない
-		if (! existsSync(out)) {
+		if (! pathExistsSync(out)) {
 			log.err += `変換失敗です。出力ファイル ${log.out} が存在しません`;
 			continue;
 		}
-		log.iSize = statSync(inp).size;
-		log.oSize = statSync(out).size;
+		log.iSize = (await stat(inp)).size;
+		log.oSize = (await stat(out)).size;
 	}
 	log_exit(0);
 });
