@@ -56,7 +56,7 @@ import {storeToRefs} from 'pinia';
 import {useField} from 'vee-validate';
 import * as yup from 'yup';
 import {openURL, on} from '../store/stVSCode';
-import {T_CFG} from '../types';
+import type {T_CFG} from '../types';
 
 const stCfg = useCfg();
 const {oCfg} = storeToRefs(stCfg);
