@@ -34,7 +34,7 @@
 		<td style="text-align: right;" v-text="e.oSize.toLocaleString('ja-JP') +' byte'"/>
 		<td v-text="(e.oSize / e.iSize).toLocaleString('ja-JP')"/>
 		<td>
-			<button type="button" id="open.readme.txt" class="btn btn-info btn-sm" @click="openURL('ws-file:///core/font/subset_font_'+ e.nm +'.txt')">Open</button>
+			<button type="button" id="open.readme.txt" class="btn btn-info btn-sm" @click="openURL(`ws-file:///${fld_src}/font/subset_font_${e.nm}.txt`)">Open</button>
 		</td>
 	</tr>
 	<tr v-if="e.err">
@@ -95,7 +95,7 @@ import {useWss, hDisabled} from '../store/stWSS';
 import {storeToRefs} from 'pinia';
 import {useOInfo} from '../store/stOInfo';
 import {ref} from 'vue';
-import {T_E2V_SELECT_ICON_INFO, T_V2E_SELECT_ICON_FILE} from '../types';
+import type {T_E2V_INIT, T_E2V_SELECT_ICON_INFO, T_V2E_SELECT_ICON_FILE} from '../types';
 
 
 const stOInfo = useOInfo();
@@ -115,13 +115,17 @@ const selectIcon = ()=> cmd2Ex(qselectIcon);
 
 const srcIcon = ref('data:image/svg+xml;base64,PHN2ZyBoZWlnaHQ9IjY0MCIgcHJlc2VydmVBc3BlY3RSYXRpbz0ieE1pZFlNaWQgbWVldCIgdmlld0JveD0iMCAwIDY0MCA2NDAiIHdpZHRoPSI2NDAiIHhtbG5zPSJodHRwOi8vd3d3LnczLm9yZy8yMDAwL3N2ZyIgeG1sbnM6eGxpbms9Imh0dHA6Ly93d3cudzMub3JnLzE5OTkveGxpbmsiPjxkZWZzPjxwYXRoIGlkPSJhIiBkPSJtMCAzMjBjMCAxNzYuNzIgMTQzLjI4IDMyMCAzMjAgMzIwczMyMC0xNDMuMjggMzIwLTMyMC0xNDMuMjgtMzIwLTMyMC0zMjAtMzIwIDE0My4yOC0zMjAgMzIwem0yMDAgMTAwdi0yMDBoODB2MjAwem0xNjAgMHYtMjAwaDgwdjIwMHoiLz48L2RlZnM+PHBhdGggZD0ibTE0Ny40OSAxODAuNDFoMzUyLjR2MjgyLjY5aC0zNTIuNHoiIGZpbGw9IiNmZmYiLz48dXNlIGZpbGw9IiMyZTJlMmUiIHhsaW5rOmhyZWY9IiNhIi8+PHVzZSBmaWxsPSJub25lIiB4bGluazpocmVmPSIjYSIvPjwvc3ZnPg==');
 const updIconImg = (src: string)=> srcIcon.value = src +'?'+ new Date().getTime();
-on('!', data=> updIconImg(data.pathIcon));
+const fld_src = ref('');
+on('!', (d: T_E2V_INIT)=> {
+	updIconImg(d.pathIcon);
+	fld_src.value = d.fld_src;
+});
 
 
 const select_icon_err = ref('');
-on('updimg', (data: T_E2V_SELECT_ICON_INFO)=> {
-	updIconImg(data.pathIcon);
-	select_icon_err.value = data.err_mes;
+on('updimg', (d: T_E2V_SELECT_ICON_INFO)=> {
+	updIconImg(d.pathIcon);
+	select_icon_err.value = d.err_mes;
 });
 
 </script>
